+++ srcs/shade.c
@@ -6,7 +6,7 @@
 /*   By: tfregni <tfregni@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/27 15:45:15 by tfregni           #+#    #+#             */
-/*   Updated: 2023/09/09 16:38:37 by tfregni          ###   ########.fr       */
+/*   Updated: 2023/09/12 16:05:44 by tfregni          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 int	cast_shadow(t_scene *scene, t_ray ray)
 {
 	int		i;
-	float	t;
+	double	t;
 
 	i = scene->shape_count - 1;
 	while (i >= 0)
@@ -31,7 +31,7 @@ int	cast_shadow(t_scene *scene, t_ray ray)
 /**
  * @param n: normal vector
  * @param p: hit point
- * @returns	a float expressing the light intensity on the point
+ * @returns	a double expressing the light intensity on the point
  * @brief if the light hits at an angle it's more diffused and
  * less intense
  * @param l: light ray (from the hit point to the light source)
@@ -46,12 +46,12 @@ int	cast_shadow(t_scene *scene, t_ray ray)
  * it by 2 to get a range of 1 and I do the same for the light
  * returned by the shadow to keep the proportion.
 */
-float	diffuse_shade(t_scene *scene, t_vector n, t_vector p)
+double	diffuse_shade(t_scene *scene, t_vector n, t_vector p)
 {
 	t_vector	l;
-	float		n_dot_l;
-	float		light;
-	float		len_l;
+	double		n_dot_l;
+	double		light;
+	double		len_l;
 	double		diff_shade;
 
 	l = vec3_sub(scene->light.pos, p);
@@ -69,7 +69,7 @@ float	diffuse_shade(t_scene *scene, t_vector n, t_vector p)
  * @param hit: hit record
  * @brief: calculates the light intensity on the hit point
 */
-float	light_coeff(t_scene *scene, t_hit_record *hit)
+double	light_coeff(t_scene *scene, t_hit_record *hit)
 {
 	t_vector	hit_pos;
 	t_vector	normal;
